+++ src/System.Private.CoreLib/shared/System/Diagnostics/Debug.cs
@@ -4,8 +4,7 @@
 
 // Do not remove this, it is needed to retain calls to these conditional methods in release builds
 #define DEBUG
-using System.Diagnostics.Contracts;
-using System.Runtime.CompilerServices;
+using System.Threading;
 
 namespace System.Diagnostics
 {
@@ -14,34 +13,39 @@ namespace System.Diagnostics
     /// </summary>
     public static partial class Debug
     {
-        private static readonly object s_lock = new object();
+        private static DebugProvider s_provider = new DebugProvider();
+
+        public static DebugProvider SetProvider(DebugProvider provider)
+        {
+            if (provider == null)
+                throw new ArgumentNullException(nameof(provider));
+
+            return Interlocked.Exchange(ref s_provider, provider);
+        }
 
         public static bool AutoFlush { get { return true; } set { } }
 
-        [ThreadStatic]
-        private static int s_indentLevel;
         public static int IndentLevel
         {
             get
             {
-                return s_indentLevel;
+                return DebugProvider.IndentLevel;
             }
             set
             {
-                s_indentLevel = value < 0 ? 0 : value;
+                DebugProvider.IndentLevel = value;
             }
         }
 
-        private static int s_indentSize = 4;
         public static int IndentSize
         {
             get
             {
-                return s_indentSize;
+                return DebugProvider.IndentSize;
             }
             set
             {
-                s_indentSize = value < 0 ? 0 : value;
+                DebugProvider.IndentSize = value;
             }
         }
 
@@ -102,7 +106,7 @@ namespace System.Diagnostics
                     stackTrace = "";
                 }
                 WriteLine(FormatAssert(stackTrace, message, detailMessage));
-                s_ShowDialog(stackTrace, message, detailMessage, "Assertion Failed");
+                s_provider.ShowDialog(stackTrace, message, detailMessage, "Assertion Failed");
             }
         }
 
@@ -120,7 +124,7 @@ namespace System.Diagnostics
                     stackTrace = "";
                 }
                 WriteLine(FormatAssert(stackTrace, message, detailMessage));
-                s_ShowDialog(stackTrace, message, detailMessage, SR.GetResourceString(failureKindMessage));
+                s_provider.ShowDialog(stackTrace, message, detailMessage, SR.GetResourceString(failureKindMessage));
             }
         }
 
@@ -138,7 +142,7 @@ namespace System.Diagnostics
 
         private static string FormatAssert(string stackTrace, string message, string detailMessage)
         {
-            string newLine = GetIndentString() + Environment.NewLine;
+            string newLine = DebugProvider.GetIndentString() + Environment.NewLine;
             return SR.DebugAssertBanner + newLine
                    + SR.DebugAssertShortMessage + newLine
                    + message + newLine
@@ -162,24 +166,7 @@ namespace System.Diagnostics
         [System.Diagnostics.Conditional("DEBUG")]
         public static void Write(string message)
         {
-            lock (s_lock)
-            {
-                if (message == null)
-                {
-                    s_WriteCore(string.Empty);
-                    return;
-                }
-                if (s_needIndent)
-                {
-                    message = GetIndentString() + message;
-                    s_needIndent = false;
-                }
-                s_WriteCore(message);
-                if (message.EndsWith(Environment.NewLine))
-                {
-                    s_needIndent = true;
-                }
-            }
+            s_provider.Write(message);
         }
 
         [System.Diagnostics.Conditional("DEBUG")]
@@ -309,42 +296,5 @@ namespace System.Diagnostics
                 WriteLine(message, category);
             }
         }
-
-        private static bool s_needIndent;
-
-        private static string s_indentString;
-
-        private static string GetIndentString()
-        {
-            int indentCount = IndentSize * IndentLevel;
-            if (s_indentString?.Length == indentCount)
-            {
-                return s_indentString;
-            }
-            return s_indentString = new string(' ', indentCount);
-        }
-
-        private sealed class DebugAssertException : Exception
-        {
-            internal DebugAssertException(string stackTrace) :
-                base(Environment.NewLine + stackTrace)
-            {
-            }
-
-            internal DebugAssertException(string message, string stackTrace) :
-                base(message + Environment.NewLine + Environment.NewLine + stackTrace)
-            {
-            }
-
-            internal DebugAssertException(string message, string detailMessage, string stackTrace) :
-                base(message + Environment.NewLine + detailMessage + Environment.NewLine + Environment.NewLine + stackTrace)
-            {
-            }
-        }
-
-        // internal and not readonly so that the tests can swap this out.
-        internal static Action<string, string, string, string> s_ShowDialog = ShowDialog;
-
-        internal static Action<string> s_WriteCore = WriteCore;
     }
 }
