+++ src/System.Private.CoreLib/shared/System/Diagnostics/DebugProvider.Unix.cs
@@ -6,11 +6,11 @@ using Microsoft.Win32.SafeHandles;
 
 namespace System.Diagnostics
 {
-    public static partial class Debug
+    public partial class DebugProvider
     {
         private static readonly bool s_shouldWriteToStdErr = Environment.GetEnvironmentVariable("COMPlus_DebugWriteToStdErr") == "1";
 
-        private static void ShowDialog(string stackTrace, string message, string detailMessage, string errorSource)
+        public virtual void ShowDialog(string stackTrace, string message, string detailMessage, string errorSource)
         {
             if (Debugger.IsAttached)
             {
